Refactor auth methods
